play-wright/requirements.txt
requests
pytest-playwright